+++ zerver/lib/event_schema.py
@@ -1075,6 +1075,7 @@ group_setting_update_data_type = DictType(
         ("create_multiuse_invite_group", int),
         ("can_access_all_users_group", int),
         ("can_create_groups", group_setting_type),
+        ("can_manage_all_groups", group_setting_type),
         ("can_create_public_channel_group", group_setting_type),
         ("can_create_private_channel_group", group_setting_type),
         ("can_create_web_public_channel_group", group_setting_type),